+++ packages/editor/src/EditorApp.tsx
@@ -22,6 +22,7 @@ import { MaterialDropZone } from "./MaterialDropZone";
 import { PerfProbe } from "./PerfProbe";
 import { ScatterPreview } from "./ScatterPreview";
 import { CityPreview } from "./CityPreview";
+import { ClipPreviewLayer } from "./ClipPreviewLayer";
 import { SelectionGizmo, ViewportSelect } from "./SelectionGizmo";
 import { TerrainReadout } from "./TerrainReadout";
 import { TerrainReadoutHud } from "./TerrainReadoutHud";
@@ -240,6 +241,7 @@ function EditorWorldOverlay({
       <TerrainSculpt api={api} ui={ui} world={world} />
       <ScatterPreview api={api} />
       <CityPreview api={api} />
+      <ClipPreviewLayer api={api} ui={ui} />
       {showGrid ? <gridHelper args={[400, 80, "#3b4252", "#20242e"]} position={[0, 0.05, 0]} /> : null}
       {readoutActive ? (
         <TerrainReadout